first pass at factoring out the schedule stuff into its own file (to avoid duplication)
